Рефакторинг: - WebView перенесен в отдельный файл - из StationSearchView вынесена логика во ViewModel
