+++ TravelSchedule/Views/CitySearchView.swift
@@ -18,51 +18,48 @@ struct CitySearchView: View {
     }
     
     var body: some View {
-        VStack {
-            
-            ScrollView {
-                if !filteredCities.isEmpty {
-                    LazyVStack(spacing: 0) {
-                        ForEach(filteredCities, id: \.self) { city in
-                            Button(action: {
-                                selectedCity = city
-                                navigation.push(.stationsList)
-                            }) {
-                                HStack {
-                                    Text(city.title ?? "(Нет названия)")
-                                        .padding(.vertical, 17)
-                                    Spacer()
-                                    Image(systemName: "chevron.forward")
-                                        .foregroundColor(.accentColor)
-                                }
-                                .padding(.horizontal)
-                                .foregroundColor(.blackDay)
+        ScrollView {
+            if !filteredCities.isEmpty {
+                LazyVStack(spacing: 0) {
+                    ForEach(filteredCities, id: \.self) { city in
+                        Button(action: {
+                            selectedCity = city
+                            navigation.push(.stationsList)
+                        }) {
+                            HStack {
+                                Text(city.title ?? "(Нет названия)")
+                                    .padding(.vertical, 17)
+                                Spacer()
+                                Image(systemName: "chevron.forward")
+                                    .foregroundColor(.accentColor)
                             }
-                            .buttonStyle(.plain)
+                            .padding(.horizontal)
+                            .foregroundColor(.blackDay)
                         }
+                        .buttonStyle(.plain)
                     }
                 }
             }
-            .searchable(
-                text: $searchText,
-                placement: .navigationBarDrawer(displayMode: .automatic),
-                prompt: Text("Введите запрос")
-            )
-            .navigationBarTitleDisplayMode(.inline)
-            .navigationTitle("Выбор города")
-            .overlay {
-                
-                if filteredCities.isEmpty && !searchText.isEmpty {
-                    Text("Город не найден")
-                        .font(.system(size: 24))
-                        .fontWeight(.bold)
-                        .foregroundColor(.blackDay)
-                }
-            }
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-            .background(Color.whiteDay)
         }
+        .searchable(
+            text: $searchText,
+            placement: .navigationBarDrawer(displayMode: .automatic),
+            prompt: Text("Введите запрос")
+        )
+        .navigationBarTitleDisplayMode(.inline)
+        .navigationTitle("Выбор города")
+        .overlay {
+            if filteredCities.isEmpty && !searchText.isEmpty {
+                Text("Город не найден")
+                    .font(.system(size: 24))
+                    .fontWeight(.bold)
+                    .foregroundColor(.blackDay)
+            }
+        }
+        .frame(maxWidth: .infinity, maxHeight: .infinity)
+        .background(Color.whiteDay)
     }
+    
 }
 
 #Preview {
