+++ TravelSchedule/Views/StationSearchView.swift
@@ -5,28 +5,11 @@ struct StationSearchView: View {
     @EnvironmentObject var stationsViewModel: StationsViewModel
     @EnvironmentObject var navigation: NavigationViewModel
     
-    @State private var searchText: String = ""
-    
-    var filteredStations: [StationModel] {
-        let allStations = stationsViewModel.allCities.flatMap { $0.stations ?? [] }
-        let filtered = searchText.isEmpty ? allStations : allStations.filter { station in
-            station.title?.lowercased().contains(searchText.lowercased()) ?? false
-        }
-        return Array(filtered.reduce(into: [String: StationModel]()) { dict, station in
-            if let yandexCode = station.codes?.yandex_code {
-                dict[yandexCode, default: station] = station
-            }
-        }.values)
-            .filter { $0.transport_type == "train" }
-            .filter { $0.title != nil && !$0.title!.isEmpty }
-            .sorted { $0.title! < $1.title! }
-    }
-    
     var body: some View {
         VStack {
             ScrollView {
                 LazyVStack(spacing: 0) {
-                    ForEach(filteredStations, id: \.self) { station in
+                    ForEach(stationsViewModel.filteredStations, id: \.self) { station in
                         Button(action: {
                             stationsViewModel.setSelectedStation(station)
                         }) {
@@ -45,14 +28,14 @@ struct StationSearchView: View {
                 }
             }
             .searchable(
-                text: $searchText,
+                text: $stationsViewModel.searchStationText,
                 placement: .navigationBarDrawer(displayMode: .automatic),
                 prompt: Text("Введите запрос")
             )
             .navigationTitle("Станции")
             .navigationBarTitleDisplayMode(.inline)
             .overlay {
-                if filteredStations.isEmpty && !searchText.isEmpty {
+                if stationsViewModel.isNoStations() {
                     Text("Станция не найдена")
                         .font(.system(size: 24))
                         .fontWeight(.bold)
